@@ -6,7 +6,11 @@
   "scripts": {
     "dev": "vite --host",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "lint": "eslint ./ --ext .tsx --ext .ts --ext .js --cache --cache-strategy content",
+    "lint:fix": "npm run lint --fix",
+    "prepare": "husky install",
+    "test": "echo \"Error: no test specified\""
   },
   "dependencies": {
     "@emotion/react": "^11.10.8",
@@ -30,11 +34,22 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.195",
+    "@types/node": "^20.5.7",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@types/validator": "^13.7.15",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
     "@vitejs/plugin-react": "^3.1.0",
-    "typescript": "^4.9.3",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-standard-with-typescript": "^39.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-n": "^16.0.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.33.2",
+    "husky": "^8.0.3",
+    "prettier": "3.0.3",
+    "typescript": "^5.2.2",
     "vite": "^4.1.0"
   },
   "packageManager": "pnpm@7.26.3",
